Mulisa-Server/requirements.txt
Django>=5,<6
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg[binary]
python-dotenv
django-filter
